bugfix : now using float instead of double

+++ textureDescriptor.cpp
@@ -57,11 +57,11 @@ std::array<cv::Mat, 8> Z(cv::Mat const& img)
 
     cv::Mat dLdx, dLdy, d2Ldx2, d2Ldy2, d2Ldxdy;
 
-    cv::Sobel(L, dLdx, CV_64F, 1, 0, CV_SCHARR);
-    cv::Sobel(L, dLdy, CV_64F, 0, 1, CV_SCHARR);
-    cv::Sobel(L, d2Ldx2, CV_64F, 2, 0);
-    cv::Sobel(L, d2Ldy2, CV_64F, 0, 2);
-    cv::Sobel(L, d2Ldxdy, CV_64F, 1, 1);
+    cv::Sobel(L, dLdx, CV_32F, 1, 0, CV_SCHARR);
+    cv::Sobel(L, dLdy, CV_32F, 0, 1, CV_SCHARR);
+    cv::Sobel(L, d2Ldx2, CV_32F, 2, 0);
+    cv::Sobel(L, d2Ldy2, CV_32F, 0, 2);
+    cv::Sobel(L, d2Ldxdy, CV_32F, 1, 1);
 
     z[0] = abs(L);
     z[1] = abs(a);
@@ -78,47 +78,47 @@ std::array<cv::Mat, 8> Z(cv::Mat const& img)
 cv::Mat Wr(int r)
 {
     int sizePatch = 2 * r + 1;
-    cv::Mat Wr(sizePatch, sizePatch, CV_64F);
-    double sigma_carre = pow((double)r / 3.0, 2);
+    cv::Mat Wr(sizePatch, sizePatch, CV_32F);
+    float sigma_carre = pow((float)r / 3.0, 2);
     if(r == 0)
         sigma_carre = 1;
-    double Z = 0.0;
+    float Z = 0.0;
     for (int i = -r; i <= r; i++) {
         for (int j = -r; j <= r; j++) {
-            double dist_pq = pow((double)sqrt(i*i + j*j), 2);
-            double w = exp(-(dist_pq) / (2.0 * sigma_carre));
-            Wr.at<double>(i + r, j + r) = w;
+            float dist_pq = pow((float)sqrt(i*i + j*j), 2);
+            float w = exp(-(dist_pq) / (2.0 * sigma_carre));
+            Wr.at<float>(i + r, j + r) = w;
             Z += w;
         }
     }
     return Wr/Z;
 }
 
-double beta(cv::Mat const& Wr)
+float beta(cv::Mat const& Wr)
 {
     //sum(Wr) is Scalar, and sum(Wr)[0] is the real part
-    return 1/(double)sum(Wr)[0];
+    return 1/(float)sum(Wr)[0];
 }
 
 inline cv::Mat Zq(std::array<cv::Mat, 8> const& Z,
         unsigned int const i,
         unsigned int const j)
 {
-    cv::Mat Zq(8,1, CV_64F);
+    cv::Mat Zq(8,1, CV_32F);
     for(unsigned int ii = 0; ii < 8; ++ii)
-        Zq.at<double>(ii) = Z[ii].at<double>(i,j);
+        Zq.at<float>(ii) = Z[ii].at<float>(i,j);
 
     return Zq;
 }
 
 inline cv::Mat MUr(std::array<cv::Mat, 8> const& Z,
         cv::Mat const& Wr,
-        double const& beta,
+        float const& beta,
         unsigned int const r,
         unsigned int const i,
         unsigned int const j)
 {
-    cv::Mat mu(8,1, CV_64F);
+    cv::Mat mu(8,1, CV_32F);
 
     //mu = beta * sum[q in N] ( Wr(p,q) * z(q) )
 
@@ -127,7 +127,7 @@ inline cv::Mat MUr(std::array<cv::Mat, 8> const& Z,
         {
             //qi in [0, 2r+1[ (therefore qi in patch width)
             //so qi-r in [-r, r]
-            mu += /*Wr.at<double>(qi, qj)* */ Zq(Z, i + qi-r, j + qj-r);
+            mu += /*Wr.at<float>(qi, qj)* */ Zq(Z, i + qi-r, j + qj-r);
         }
 
     //mu *= beta;
@@ -136,10 +136,10 @@ inline cv::Mat MUr(std::array<cv::Mat, 8> const& Z,
     return mu;
 }
 
-std::vector<std::vector<cv::Mat> > Crp(std::array<cv::Mat, 8> const& Z, cv::Mat const& Wr, double const& beta, unsigned int r)
+std::vector<std::vector<cv::Mat> > Crp(std::array<cv::Mat, 8> const& Z, cv::Mat const& Wr, float const& beta, unsigned int r)
 {
-    std::cout << "Z : " << Z[0].at<double>(0,0) << std::endl;
-    std::cout << "Wr : " << Wr.at<double>(0,0) << std::endl;
+    std::cout << "Z : " << Z[0].at<float>(0,0) << std::endl;
+    std::cout << "Wr : " << Wr.at<float>(0,0) << std::endl;
     std::cout << "beta : " << beta << std::endl;
     std::cout << "r : " << r << std::endl;
 
@@ -156,16 +156,18 @@ std::vector<std::vector<cv::Mat> > Crp(std::array<cv::Mat, 8> const& Z, cv::Mat 
             cv::Mat mu = MUr(Z, Wr, beta, r, i, j);
             //std::cout << mu << std::endl;
 
-            cv::Mat Crp = cv::Mat::zeros(8,8, CV_64F);
+            cv::Mat Crp = cv::Mat::zeros(8,8, CV_32F);
 
             for(unsigned int qi = 0; qi < 2*r+1; ++qi)
                 for(unsigned int qj = 0; qj < 2*r+1; ++qj)
                 {
                     cv::Mat zq = Zq(Z, i + qi-r, j + qj-r) - mu;
-                    Crp += (zq*zq.t())*(Wr.at<double>(qi,qj));
+                    //std::cout << "Zq [" << qi << "," << qj << "]" << std::endl;
+                    //std::cout << zq << std::endl;
+                    Crp += (zq*zq.t())*(Wr.at<float>(qi,qj));
                 }
 
-            //std::cout << Crp.at<double>(0, 0) << std::endl;
+            //std::cout << Crp.at<float>(0, 0) << std::endl;
 
             Crp *= beta;
